Add DTO classes
